backend/app/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.0.1
beautifulsoup4==4.12.3
botocore==1.34.115
bs4==0.0.2
build==1.2.1
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
colorama==0.4.6
crashtest==0.4.1
cryptography==42.0.5
distlib==0.3.8
distro==1.9.0
dnspython==2.6.1
dulwich==0.21.7
email_validator==2.1.1
fastapi==0.110.1
fastjsonschema==2.19.1
filelock==3.14.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
jmespath==1.0.1
joblib==1.4.2
keyring==24.3.1
more-itertools==10.2.0
msgpack==1.0.8
numpy==1.26.4
openai==1.16.1
packaging==24.0
pandas==2.2.2
passlib==1.7.4
pexpect==4.9.0
pip-tools==7.4.1
pkginfo==1.10.0
platformdirs==4.2.2
pluggy==1.4.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.7.2
pyparsing==3.1.2
pyproject_hooks==1.1.0
pytest==8.1.1
pytest-asyncio==0.23.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
pywin32-ctypes==0.2.2
rapidfuzz==3.9.1
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9
s3transfer==0.10.1
scikit-learn==1.5.0
scipy==1.13.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.37.2
threadpoolctl==3.5.0
tomlkit==0.12.5
tqdm==4.66.2
trove-classifiers==2024.5.22
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
virtualenv==20.26.2
xmltodict==0.13.0
zipp==3.19.2
